refactor(merge): split commit + adoption + survivor family into MergeDialog.Commit.cs
